Change __init__.py syntax and add mps_utils
mps_utils is a module for various functionality for MPS/MPO.
