+++ tncontract/onedim/__init__.py
@@ -4,15 +4,10 @@ onedimensional
 
 Subpackage for one dimensional tensor networks
 """
-from tncontract.onedim.one_dimension import (MatrixProductState, 
-        MatrixProductOperator,
-        OneDimensionalTensorNetwork, check_canonical_form_mps,
-        contract_mps_mpo, contract_multi_index_tensor_with_one_dim_array,
-        contract_virtual_indices, frob_distance_squared, inner_product_mps,
-        inner_product_one_dimension, left_canonical_form_mps,
-        mps_complex_conjugate, onebody_sum_mpo, reverse_mps,
-        right_canonical_form_mps, svd_compress_mps, variational_compress_mps)
+from tncontract.onedim.one_dimension import *
+import tncontract.onedim.mps_methods
 
-from tncontract.onedim.one_dimension import(init_mps_random, 
-        tensor_to_mps, tensor_to_mpo, expvals_mps)
+
+#from tncontract.onedim.one_dimension import(init_mps_random, 
+#        tensor_to_mps, tensor_to_mpo, expvals_mps, onebody_sum_mpo)
